+++ PostHog/PostHogSDK.swift
@@ -46,6 +46,7 @@ let maxRetryDelay = 30.0
     private var capturedAppInstalled = false
     private var appFromBackground = false
     private var isInBackground = false
+    private var isiOSNativeSdk = true
     #if os(iOS)
         private var replayIntegration: PostHogReplayIntegration?
     #endif
@@ -143,6 +144,9 @@ let maxRetryDelay = 30.0
 
             registerNotifications()
             captureScreenViews()
+            
+            // postHogSdkName will be set to eg posthog-react-native if not
+            isiOSNativeSdk = postHogSdkName == postHogiOSSdkName
 
             PostHogSessionManager.shared.startSession()
 
@@ -597,7 +601,8 @@ let maxRetryDelay = 30.0
         }
 
         // If events fire in the background after the threshold, they should no longer have a sessionId
-        if isInBackground {
+        // for hybrid SDKs, the session is handled by the hybrid SDK
+        if isInBackground, isiOSNativeSdk {
             PostHogSessionManager.shared.resetSessionIfExpired {
                 self.resetViews()
             }
@@ -1150,8 +1155,11 @@ let maxRetryDelay = 30.0
     }
 
     @objc func handleAppDidBecomeActive() {
-        PostHogSessionManager.shared.rotateSessionIdIfRequired {
-            self.resetViews()
+        // for hybrid SDKs, the session is handled by the hybrid SDK
+        if isiOSNativeSdk {
+            PostHogSessionManager.shared.rotateSessionIdIfRequired {
+                self.resetViews()
+            }
         }
 
         isInBackground = false
@@ -1188,7 +1196,10 @@ let maxRetryDelay = 30.0
     @objc func handleAppDidEnterBackground() {
         captureAppBackgrounded()
 
-        PostHogSessionManager.shared.updateSessionLastTime()
+        // for hybrid SDKs, the session is handled by the hybrid SDK
+        if isiOSNativeSdk {
+            PostHogSessionManager.shared.updateSessionLastTime()
+        }
 
         isInBackground = true
     }