+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- recording: do not rotate the session id for hybrid SDKs ([#249](https://github.com/PostHog/posthog-ios/pull/249))
+
 ## 3.15.2 - 2024-11-13
 
 - fix: allow changing person properties after identify ([#249](https://github.com/PostHog/posthog-ios/pull/249))